# business-central-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Business Central Backend API
  version: 0.1.0
  description: |
    Contract-first OpenAPI documentation for the Business Central backend.

    Health, authentication, authorization, commerce, reporting, service, and
    repair endpoints are implemented. Reusable schemas, security conventions,
    response envelopes, required error codes, and collection query semantics
    are documented here.
  license:
    name: Proprietary
  x-business-central-status: implemented-current-scope
servers:
  - url: http://localhost:8080
    description: Local backend (when the HTTP runtime is available)
...
paths:
...
  /api/v1/reports/sales-summary:
    get:
      tags: [Reports]
      summary: Get sales and profit summary
      operationId: getSalesSummary
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Sales totals, refunds, cost of goods sold, and gross profit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalesSummaryEnvelope' }
  /api/v1/reports/profit-summary:
    get:
      tags: [Reports]
      summary: Get profit summary
      operationId: getProfitSummary
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Profit summary using canonical sales and inventory costs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalesSummaryEnvelope' }
  /api/v1/reports/sales-by-day:
    get:
      tags: [Reports]
      summary: Analyze sales by UTC day
      operationId: listSalesByDay
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated daily sales and profit analysis.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalesByDayCollectionEnvelope' }
  /api/v1/reports/top-products:
    get:
      tags: [Reports]
      summary: Analyze top-selling products
      operationId: listTopProducts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/PageIndex'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated product sales, cost, and profit analysis.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopProductCollectionEnvelope' }
...
  /api/v1/sync/handshake:
    post:
      tags: [Synchronization]
      summary: Register a client device and open a synchronization session
      operationId: synchronizationHandshake
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncHandshakeRequest' }
      responses:
        '200':
          description: Synchronization protocol and session information.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncHandshakeEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/sync/push:
    post:
      tags: [Synchronization]
      summary: Upload idempotent client operations
      description: Tenant-scoped settings, catalog, measurement, delivery, repair, and POS synchronization operations are accepted according to their domain policies. Repair aggregate REPAIR_TICKET CREATE is atomic across all work items and preserves supplied work-item UUIDs. REPAIR_DIAGNOSTIC, REPAIR_IMAGE, REPAIR_PART, REPAIR_APPROVAL, REPAIR_WARRANTY, and cash-only REPAIR_PAYMENT CREATE operations are dependency-ordered behind the parent and validate tenant, shop, repair-order, and optional work-item scope. Catalog-part promotions and external payment methods remain online-only; unsupported mutations return visible rejection results.
      operationId: synchronizationPush
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncPushRequest' }
      responses:
        '200':
          description: Per-operation results, including accepted, rejected, and conflict outcomes.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncPushEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/sync/conflicts/{operationId}/resolve:
    post:
      tags: [Synchronization]
      summary: Resolve an open synchronization conflict
      description: SHOP_SETTINGS conflicts can keep the authoritative server copy or explicitly rebase and apply the validated client change. Resolution is tenant-scoped and idempotent.
      operationId: resolveSynchronizationConflict
      security: [{ bearerAuth: [] }]
      parameters:
        - name: operationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncConflictResolutionRequest' }
      responses:
        '200':
          description: Authoritative conflict resolution result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncConflictResolutionEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/sync/pull:
    post:
      tags: [Synchronization]
      summary: Download merchant changes after a durable checkpoint
      operationId: synchronizationPull
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncPullRequest' }
      responses:
        '200':
          description: Ordered synchronization changes and checkpoint information.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SyncPullEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send the access token returned by login or refresh as
        Authorization: Bearer <access-token>. Refresh tokens are rotated by
        POST /api/v1/auth/refresh.
  parameters:
...
    Query:
      name: query
      in: query
      required: false
      description: Case-insensitive text search over the endpoint's searchable fields.
      schema: { type: string }
    Filter:
      name: filter
      in: query
      required: false
      description: Comma- or semicolon-separated equality filters, for example status:OPEN,is_active:true.
      schema: { type: string }
    PageIndex:
      name: page_index
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    BadRequest:
      description: The request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              $ref: '#/components/examples/ValidationError'
    Unauthorized:
      description: Authentication is required or the credentials are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist in the authorized tenant context.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The command conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: UUID identifier.
...
    ResourceMeta:
      type: object
      description: Optional metadata for a single-resource response.
      additionalProperties: true
    CollectionMeta:
      type: object
      required:
        - page
        - page_size
        - total
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
      additionalProperties: false
    ResourceEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: object
          description: Concrete endpoint resource payload.
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/ResourceMeta'
      additionalProperties: false
...
    SyncHandshakeRequest:
      type: object
      required: [device_identifier, client_session_key]
      properties:
        device_identifier: { type: string, minLength: 1, maxLength: 255 }
        device_name: { type: string, maxLength: 255 }
        client_session_key: { type: string, minLength: 1, maxLength: 255 }
        scope: { type: string, default: merchant, maxLength: 100 }
    SyncHandshakeEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResourceEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SyncHandshake'
    SyncHandshake:
      type: object
      required: [protocol_version, schema_version, device, session, server_sequence]
      properties:
        protocol_version: { type: string }
        schema_version: { type: string }
        server_sequence: { type: integer, format: int64, minimum: 0 }
        device: { $ref: '#/components/schemas/SyncDevice' }
        session: { $ref: '#/components/schemas/SyncSession' }
    SyncDevice:
      type: object
      required: [id, merchant_id, device_identifier, last_seen_at, is_active]
      properties:
        id: { type: string, format: uuid }
        merchant_id: { type: string, format: uuid }
        membership_id: { type: string, format: uuid, nullable: true }
        device_identifier: { type: string }
        device_name: { type: string, nullable: true }
        last_seen_at: { type: string, format: date-time }
        is_active: { type: boolean }
    SyncSession:
      type: object
      required: [id, client_session_key, status, scope, last_server_sequence, started_at]
      properties:
        id: { type: string, format: uuid }
        client_session_key: { type: string }
        status: { type: string, enum: [OPEN, COMPLETED, FAILED, CANCELLED] }
        scope: { type: string }
        last_server_sequence: { type: integer, format: int64, minimum: 0 }
        started_at: { type: string, format: date-time }
    SyncOperation:
      type: object
      required: [operation_id, entity_type, entity_id, operation_type, payload]
      properties:
        operation_id: { type: string, maxLength: 255 }
        entity_type: { type: string, maxLength: 100 }
        entity_id: { type: string, format: uuid }
        shop_id: { type: string, format: uuid, nullable: true }
        operation_type: { type: string, enum: [CREATE, UPDATE, DELETE] }
        base_version: { type: integer, format: int64, minimum: 0, nullable: true }
        payload_hash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: Optional canonical JSON SHA-256 integrity value; the server verifies it when supplied.
        dependency_operation_id:
          type: string
          maxLength: 255
          description: Client operation identifier that must already be applied on the same tenant-scoped device.
        client_created_at: { type: string, format: date-time, nullable: true }
        payload: { type: object, additionalProperties: true }
    SyncPushRequest:
      type: object
      required: [session_id, operations]
      properties:
        session_id: { type: string, format: uuid }
        operations:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/SyncOperation' }
    SyncOperationResult:
      type: object
      required: [operation_id, status]
      properties:
        operation_id: { type: string }
        server_operation_id: { type: string, format: uuid }
        status: { type: string, enum: [APPLIED, REJECTED, CONFLICT] }
        code: { type: string }
        message: { type: string }
        server_sequence: { type: integer, format: int64, nullable: true }
        entity_version: { type: integer, format: int64, nullable: true }
        server_payload: { type: object, additionalProperties: true, nullable: true }
    SyncPushResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/SyncOperationResult' }
    SyncPushEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResourceEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SyncPushResponse' }
    SyncConflictResolutionRequest:
      type: object
      required: [strategy]
      properties:
        strategy: { type: string, enum: [KEEP_SERVER, APPLY_CLIENT] }
    SyncConflictResolution:
      type: object
      required: [operation_id, strategy, status, entity_version, server_payload]
      properties:
        operation_id: { type: string, format: uuid }
        strategy: { type: string, enum: [KEEP_SERVER, APPLY_CLIENT] }
        status: { type: string, enum: [IGNORED, RESOLVED] }
        entity_version: { type: integer, format: int64, minimum: 0 }
        server_sequence: { type: integer, format: int64, nullable: true }
        server_payload: { type: object, additionalProperties: true }
    SyncConflictResolutionEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResourceEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SyncConflictResolution' }
    SyncPullRequest:
      type: object
      required: [session_id, after_sequence]
      properties:
        session_id: { type: string, format: uuid }
        scope: { type: string, default: merchant }
        after_sequence: { type: integer, format: int64, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 500, default: 100 }
    SyncChange:
      type: object
      required: [server_sequence, entity_type, entity_id, entity_version, operation_type, payload, created_at]
      properties:
        server_sequence: { type: integer, format: int64 }
        entity_type: { type: string }
        entity_id: { type: string, format: uuid }
        entity_version: { type: integer, format: int64 }
        operation_type: { type: string, enum: [CREATE, UPDATE, DELETE] }
        operation_id: { type: string, format: uuid, nullable: true }
        payload: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    SyncPullResponse:
      type: object
      required: [scope, changes, next_sequence, current_sequence, has_more]
      properties:
        scope: { type: string }
        changes: { type: array, items: { $ref: '#/components/schemas/SyncChange' } }
        next_sequence: { type: integer, format: int64 }
        current_sequence: { type: integer, format: int64 }
        has_more: { type: boolean }
    SyncPullEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResourceEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SyncPullResponse' }
    SalesSummary:
      type: object
      properties:
        from: { type: string, format: date-time, nullable: true }
        to: { type: string, format: date-time, nullable: true }
        order_count: { type: integer, format: int64 }
        pos_order_count: { type: integer, format: int64 }
        repair_count: { type: integer, format: int64 }
        item_quantity: { type: string }
        gross_sales: { type: string }
        discounts: { type: string }
        tax: { type: string }
        net_sales: { type: string }
        refunds: { type: string }
        cost_of_goods_sold: { type: string }
        gross_profit: { type: string }
        gross_margin_percent: { type: string }
    SalesSummaryEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/SalesSummary' }
        meta: { $ref: '#/components/schemas/ResourceMeta' }
    SalesByDay:
      type: object
      properties:
        day: { type: string, format: date-time }
        order_count: { type: integer, format: int64 }
        item_quantity: { type: string }
        net_sales: { type: string }
        refunds: { type: string }
        cost_of_goods_sold: { type: string }
        gross_profit: { type: string }
    SalesByDayCollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/SalesByDay' } }
        meta: { $ref: '#/components/schemas/CollectionMeta' }
    TopProduct:
      type: object
      properties:
        product_id: { $ref: '#/components/schemas/UUID' }
        variant_id: { $ref: '#/components/schemas/UUID' }
        product_name: { type: string }
        variant_name: { type: string }
        sku: { type: string }
        item_quantity: { type: string }
        net_sales: { type: string }
        cost_of_goods_sold: { type: string }
        gross_profit: { type: string }
    TopProductCollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/TopProduct' } }
        meta: { $ref: '#/components/schemas/CollectionMeta' }
...
    ErrorFields:
      type: object
      description: Field-level validation details keyed by request field name.
      additionalProperties:
        oneOf:
          - type: string
          - type: array
            items:
              type: string
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_CONFLICT
            - MODULE_DISABLED
            - OFFLINE_OPERATION_REJECTED
            - INTERNAL_ERROR
        message:
          type: string
        fields:
          $ref: '#/components/schemas/ErrorFields'
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
        - error
        - request_id
      properties:
        error:
          $ref: '#/components/schemas/Error'
        request_id:
          $ref: '#/components/schemas/UUID'
      additionalProperties: false
  examples:
    ValidationError:
      summary: Validation failure
      value:
        error:
          code: VALIDATION_ERROR
          message: The request is invalid.
          fields:
            name: Name is required.
        request_id: 00000000-0000-0000-0000-000000000000